+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 * /feed design improved
 * Redirect to /account after register so they can change their profiles
 * Newsfeed comes with individual feed links
+
+## 0.3 to 0.4
+* feed permalinks now show in "dialog" not window.alert
+* member management